Made constructor package private and renamed test class
