+++ src/main/java/zowe/client/sdk/zosconsole/method/IssueConsole.java
@@ -61,7 +61,7 @@ public class IssueConsole {
      * @param request    any compatible ZoweRequest Interface object
      * @author Frank Giordano
      */
-    public IssueConsole(final ZosConnection connection, final ZosmfRequest request) {
+    IssueConsole(final ZosConnection connection, final ZosmfRequest request) {
         ValidateUtils.checkIllegalParameter(connection == null, "connection is null");
         ValidateUtils.checkNullParameter(request == null, "request is null");
         this.connection = connection;
