+++ src/main/javadoc/zowe/client/sdk/zosconsole/IssueCommandTest.html
@@ -123,7 +123,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>zowe.client.sdk.zosconsole.IssueCommandTest</li>
+<li>zowe.client.sdk.zosconsole.method.IssueConsoleTest</li>
 </ul>
 </li>
 </ul>
